+++ packages/lumx-react/src/components/tabs/Tab.test.tsx
@@ -10,7 +10,9 @@ import { CLASSNAME, Tab, TabProps } from './Tab';
 import { setupTabProviderMocks } from './test.mocks';
 
 // Mock useTabProviderContext.
-jest.mock('./state', () => ({ useTabProviderContext: jest.fn() }));
+jest.mock('./state', () => {
+    return { useTabProviderContext: jest.fn(), useTabProviderContextState: jest.fn() };
+});
 
 /**
  * Define the overriding properties waited by the `setup` function.