+++ packages/site-demo/content/product/components/progress-tracker/react/default.tsx
@@ -1,28 +1,49 @@
-import { ProgressTracker, ProgressTrackerStep } from '@lumx/react';
+import { ProgressTracker, ProgressTrackerProvider, ProgressTrackerStep } from '@lumx/react';
 import React, { useState } from 'react';
 
-export const App = ({ theme }: any) => {
-    const [activeIndex, setActiveIndex] = useState(2);
-    const stepProps = [
-        { isComplete: true },
-        { isComplete: true },
-        { hasError: true },
-        { isComplete: false },
-        { isComplete: false },
+export const App = () => {
+    const [activeStep, setActiveStep] = useState(2);
+
+    const steps = [
+        {
+            isDisabled: false,
+            isComplete: true,
+            hasError: false,
+        },
+        {
+            isDisabled: false,
+            isComplete: true,
+            hasError: false,
+        },
+        {
+            isDisabled: false,
+            isComplete: false,
+            hasError: true,
+        },
+        {
+            isDisabled: true,
+            isComplete: false,
+            hasError: false,
+        },
+        {
+            isDisabled: true,
+            isComplete: false,
+            hasError: false,
+        },
     ];
 
     return (
-        <ProgressTracker activeStep={activeIndex} theme={theme}>
-            {stepProps.map((step, index) => (
-                <ProgressTrackerStep
-                    key={index}
-                    isActive={activeIndex === index}
-                    onClick={index < 3 ? () => setActiveIndex(index) : null}
-                    label={`Step ${index + 1}`}
-                    helper={index === 2 ? 'Error message' : `Helper text ${index + 1}`}
-                    {...step}
-                />
-            ))}
-        </ProgressTracker>
+        <ProgressTrackerProvider activeStepIndex={activeStep} onChange={setActiveStep}>
+            <ProgressTracker aria-label="Steps with a linear progression">
+                {steps.map((step, index) => (
+                    <ProgressTrackerStep
+                        key={`step-${index + 1}`}
+                        label={`Step ${index + 1}`}
+                        helper={index === 2 ? 'Error message' : `Helper text ${index + 1}`}
+                        {...step}
+                    />
+                ))}
+            </ProgressTracker>
+        </ProgressTrackerProvider>
     );
 };